Add EXPOSE to Dockerfile
You want to ensure it's "documented" that the container wants to open that port.

While it is just "a hint to the next person to maintain the system" so that they know what are the ports should be published when runs the image, it's helpful. EXPOSE doesn't publish it, it just lets docker inspect know the container WANTS to. Helpful for humans, and software which automate containers.

In practicality, when running the image with -P, the exposed ports in dockerfile will be published to avoid a long list of -p for individual ports.